pyproject.toml
[tool.poetry]
name = "fallen_london_chronicler"
version = "1.0.0"
description = "A tool for recording and exporting Fallen London content."
authors = ["Lyrositor"]

[tool.poetry.dependencies]
python = ">=3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
SQLAlchemy = "^1.3.22"
itsdangerous = "^1.1.0"
aiofiles = "^0.6.0"
Jinja2 = "^2.11.2"
aiohttp = "^3.7.3"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.15.0"
